+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 ## [Unreleased]
 
 ### Added
+- Automatic and manual pipeline execution mode (@mkasztelnik)
 
 ### Changed